requirements.txt
alpaca-py==0.19.1
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==11.0.3
